+++ marshmallow/compat.py
@@ -5,7 +5,6 @@ import functools
 import inspect
 
 PY2 = int(sys.version_info[0]) == 2
-PY26 = PY2 and int(sys.version_info[1]) < 7
 
 if PY2:
     import urlparse
@@ -19,18 +18,6 @@ if PY2:
     itervalues = lambda d: d.itervalues()
     iteritems = lambda d: d.iteritems()
     zip_longest = itertools.izip_longest
-    if PY26:
-        from .ordereddict import OrderedDict
-    else:
-        from collections import OrderedDict
-    OrderedDict = OrderedDict
-    def get_func_args(func):
-        if isinstance(func, functools.partial):
-            return list(inspect.getargspec(func.func).args)
-        if inspect.isfunction(func) or inspect.ismethod(func):
-            return list(inspect.getargspec(func).args)
-        if callable(func):
-            return list(inspect.getargspec(func.__call__).args)
 else:
     import urllib.parse
     urlparse = urllib.parse
@@ -43,16 +30,6 @@ else:
     itervalues = lambda d: d.values()
     iteritems = lambda d: d.items()
     zip_longest = itertools.zip_longest
-    from collections import OrderedDict
-    OrderedDict = OrderedDict
-    def get_func_args(func):
-        if isinstance(func, functools.partial):
-            return list(inspect.signature(func.func).parameters)
-        if inspect.isfunction(func):
-            return list(inspect.signature(func).parameters)
-        if callable(func) or inspect.ismethod(func):
-            return ['self'] + list(inspect.signature(func.__call__).parameters)
-
 
 # From six
 def with_metaclass(meta, *bases):