+++ tests/test_schema.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import datetime as dt
 import simplejson as json
 import decimal
 import random
-from collections import namedtuple
+from collections import namedtuple, OrderedDict
 
 import pytest
 
@@ -13,7 +14,6 @@ from marshmallow import (
     validates, validates_schema
 )
 from marshmallow.exceptions import ValidationError
-from marshmallow.compat import OrderedDict
 
 from tests.base import *  # noqa
 
@@ -86,6 +86,57 @@ def test_load_resets_errors():
     result = schema.load({'name': 'Joe', 'email': '__invalid'})
     assert len(result.errors['email']) == 1
     assert 'Not a valid email address.' in result.errors['email'][0]
+
+def test_strict_load_validation_error_stores_input_data_and_valid_data():
+    class MySchema(Schema):
+        always_valid = fields.DateTime()
+        always_invalid = fields.Field(validate=[lambda v: False])
+
+        class Meta:
+            strict = True
+
+    schema = MySchema()
+    input_data = {'always_valid': dt.datetime.utcnow().isoformat(), 'always_invalid': 24}
+    try:
+        schema.load(input_data)
+    except ValidationError as err:
+        # err.data is the raw input data
+        assert err.data == input_data
+        assert 'always_valid' in err.valid_data
+        # err.valid_data contains valid, deserialized data
+        assert isinstance(err.valid_data['always_valid'], dt.datetime)
+        # excludes invalid data
+        assert 'always_invalid' not in err.valid_data
+    else:
+        pytest.fail('Data is invalid. Expected a ValidationError to be raised.')
+
+def test_strict_dump_validation_error_stores_partially_valid_data():
+    class FailOnDump(fields.Field):
+        def _serialize(self, *args, **kwargs):
+            raise ValidationError('fail')
+
+    class MySchema(Schema):
+        always_valid = fields.DateTime()
+        always_invalid = FailOnDump()
+
+        class Meta:
+            strict = True
+
+    schema = MySchema()
+    input_data = {'always_valid': dt.datetime.utcnow(), 'always_invalid': 24}
+    try:
+        schema.dump(input_data)
+    except ValidationError as err:
+        # err.data is the raw input data
+        assert err.data == input_data
+        assert 'always_valid' in err.valid_data
+        # err.valid_data contains valid, serialized data
+        assert isinstance(err.valid_data['always_valid'], str)
+        # excludes invalid data
+        assert 'always_invalid' not in err.valid_data
+    else:
+        pytest.fail('Data is invalid. Expected a ValidationError to be raised.')
+
 
 def test_dump_resets_error_fields():
     class MySchema(Schema):
@@ -241,9 +292,8 @@ def test_dumps_returns_a_marshalresult(user):
     assert type(result.data) == str
     assert type(result.errors) == dict
 
-def test_dumping_single_object_with_collection_schema():
+def test_dumping_single_object_with_collection_schema(user):
     s = UserSchema(many=True)
-    user = UserSchema('Mick')
     result = s.dump(user, many=False)
     assert type(result.data) == dict
     assert result.data == UserSchema().dump(user).data
@@ -410,8 +460,8 @@ class TestValidate:
 @pytest.mark.parametrize('SchemaClass',
     [UserSchema, UserMetaSchema])
 def test_fields_are_not_copies(SchemaClass):
-    s = SchemaClass(User('Monty', age=42))
-    s2 = SchemaClass(User('Monty', age=43))
+    s = SchemaClass()
+    s2 = SchemaClass()
     assert s.fields is not s2.fields
 
 
@@ -509,16 +559,6 @@ def test_as_string():
     serialized = UserFloatStringSchema().dump(u)
     assert type(serialized.data['age']) == str
     assert_almost_equal(float(serialized.data['age']), 42.3)
-
-def test_extra():
-    user = User("Joe", email="joe@foo.com")
-    data, errors = UserSchema(extra={"fav_color": "blue"}).dump(user)
-    assert data['fav_color'] == "blue"
-
-def test_extra_many():
-    users = [User('Fred'), User('Brian')]
-    data, errs = UserSchema(many=True, extra={'band': 'Queen'}).dump(users)
-    assert data[0]['band'] == 'Queen'
 
 @pytest.mark.parametrize('SchemaClass',
     [UserSchema, UserMetaSchema])
@@ -633,18 +673,30 @@ def test_invalid_dict_but_okay():
     s = UserSchema().dump(u)
     assert 'various_data' not in s.errors
 
-def test_custom_json():
-    class UserJSONSchema(Schema):
-        name = fields.String()
+def test_json_module_is_deprecated():
+    with pytest.warns(DeprecationWarning):
+        class UserJSONSchema(Schema):
+            name = fields.String()
 
-        class Meta:
-            json_module = mockjson
+            class Meta:
+                json_module = mockjson
 
     user = User('Joe')
     s = UserJSONSchema()
     result, errors = s.dumps(user)
     assert result == mockjson.dumps('val')
 
+def test_render_module():
+    class UserJSONSchema(Schema):
+        name = fields.String()
+
+        class Meta:
+            render_module = mockjson
+
+    user = User('Joe')
+    s = UserJSONSchema()
+    result, errors = s.dumps(user)
+    assert result == mockjson.dumps('val')
 
 def test_custom_error_message():
     class ErrorSchema(Schema):
@@ -699,6 +751,7 @@ def test_nested_only():
         foo = fields.Field()
         bar = fields.Field()
         baz = fields.Field()
+
     class ParentSchema(Schema):
         bla = fields.Field()
         bli = fields.Field()
@@ -720,6 +773,7 @@ def test_nested_exclude():
         foo = fields.Field()
         bar = fields.Field()
         baz = fields.Field()
+
     class ParentSchema(Schema):
         bla = fields.Field()
         bli = fields.Field()
@@ -736,11 +790,12 @@ def test_nested_exclude():
     assert 'baz' not in child
 
 
-def test_nested_only_and_exclude():
+def test_nested_only_and_exclude_with_dot_notation():
     class ChildSchema(Schema):
         foo = fields.Field()
         bar = fields.Field()
         baz = fields.Field()
+
     class ParentSchema(Schema):
         bla = fields.Field()
         bli = fields.Field()
@@ -762,10 +817,12 @@ def test_meta_nested_exclude():
         foo = fields.Field()
         bar = fields.Field()
         baz = fields.Field()
+
     class ParentSchema(Schema):
         bla = fields.Field()
         bli = fields.Field()
         blubb = fields.Nested(ChildSchema)
+
         class Meta:
             exclude = ('blubb.foo',)
     sch = ParentSchema()
@@ -785,10 +842,12 @@ def test_deeply_nested_only_and_exclude():
         goo = fields.Field()
         gah = fields.Field()
         bah = fields.Field()
+
     class ChildSchema(Schema):
         foo = fields.Field()
         bar = fields.Field()
         flubb = fields.Nested(GrandChildSchema)
+
     class ParentSchema(Schema):
         bla = fields.Field()
         bli = fields.Field()
@@ -948,10 +1007,12 @@ def test_nested_constructor_only_and_exclude():
         goo = fields.Field()
         gah = fields.Field()
         bah = fields.Field()
+
     class ChildSchema(Schema):
         foo = fields.Field()
         bar = fields.Field()
         flubb = fields.Nested(GrandChildSchema)
+
     class ParentSchema(Schema):
         bla = fields.Field()
         bli = fields.Field()
@@ -1184,19 +1245,7 @@ class MySchema(Schema):
         raise CustomError('Something bad happened')
 
 
-class TestErrorHandler:
-
-    def test_error_handler_decorator_is_deprecated(self):
-
-        def deprecated():
-            class MySchema(Schema):
-                pass
-
-            @MySchema.error_handler
-            def f(*args, **kwargs):
-                pass
-
-        pytest.deprecated_call(deprecated)
+class TestHandleError:
 
     def test_dump_with_custom_error_handler(self, user):
         user.age = 'notavalidage'
@@ -1721,7 +1770,7 @@ class TestSelfReference:
 
         assert errors == {
             'basic': {
-                'sub_basics':  [u'Missing data for required field.'] ,
+                'sub_basics': [u'Missing data for required field.'],
                 'simple_field': [u'Missing data for required field.'],
             }
         }
@@ -1876,6 +1925,23 @@ class TestContext:
         msg = 'No context available for Function field {0!r}'.format('is_collab')
         assert msg in str(excinfo)
 
+    def test_function_field_handles_bound_serializer(self):
+        class SerializeA(object):
+            def __call__(self, value):
+                return 'value'
+        serialize = SerializeA()
+
+        # only has a function field
+        class UserFunctionContextSchema(Schema):
+            is_collab = fields.Function(serialize)
+
+        owner = User('Joe')
+        serializer = UserFunctionContextSchema(strict=True)
+        # no context
+        serializer.context = None
+        data = serializer.dump(owner)[0]
+        assert data['is_collab'] is 'value'
+
     def test_fields_context(self):
         class CSchema(Schema):
             name = fields.String()
@@ -1964,30 +2030,18 @@ class TestFieldInheritance:
             field_d = expected['field_d']
         assert SerializerD._declared_fields == expected
 
-def get_from_dict(schema, key, obj, default=None):
+def get_from_dict(schema, obj, key, default=None):
     return obj.get('_' + key, default)
 
-class TestAccessor:
+class TestGetAttribute:
 
-    def test_accessor_decorator_is_deprecated(self):
-
-        def deprecated():
-            class MySchema(Schema):
-                pass
-
-            @MySchema.accessor
-            def f(*args, **kwargs):
-                pass
-
-        pytest.deprecated_call(deprecated)
-
-    def test_accessor_is_used(self):
+    def test_get_attribute_is_used(self):
         class UserDictSchema(Schema):
             name = fields.Str()
             email = fields.Email()
 
-            def get_attribute(self, attr, obj, default):
-                return get_from_dict(self, attr, obj, default)
+            def get_attribute(self, obj, attr, default):
+                return get_from_dict(self, obj, attr, default)
 
         user_dict = {'_name': 'joe', '_email': 'joe@shmoe.com'}
         schema = UserDictSchema()
@@ -2000,13 +2054,13 @@ class TestAccessor:
         with pytest.raises(AttributeError):
             schema.dump(user)
 
-    def test_accessor_with_many(self):
+    def test_get_attribute_with_many(self):
         class UserDictSchema(Schema):
             name = fields.Str()
             email = fields.Email()
 
-            def get_attribute(self, attr, obj, default):
-                return get_from_dict(self, attr, obj, default)
+            def get_attribute(self, obj, attr, default):
+                return get_from_dict(self, obj, attr, default)
 
         user_dicts = [{'_name': 'joe', '_email': 'joe@shmoe.com'},
                       {'_name': 'jane', '_email': 'jane@shmane.com'}]