+++ tests/test_serialization.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Tests for field serialization."""
-from collections import namedtuple
+from collections import namedtuple, OrderedDict
 import datetime as dt
 import itertools
 import decimal
@@ -10,7 +10,7 @@ import pytest
 
 from marshmallow import Schema, fields, utils
 from marshmallow.exceptions import ValidationError
-from marshmallow.compat import basestring, OrderedDict
+from marshmallow.compat import basestring
 from marshmallow.utils import missing as missing_
 
 from tests.base import User, ALL_FIELDS
@@ -62,12 +62,32 @@ class TestFieldSerialization:
         field = fields.Function(lambda obj: obj.name.upper())
         assert "FOO" == field.serialize("key", user)
 
+    def test_function_field_passed_serialize_only_is_dump_only(self, user):
+        field = fields.Function(serialize=lambda obj: obj.name.upper())
+        assert field.dump_only is True
+
+    def test_function_field_passed_deserialize_and_serialize_is_not_dump_only(self):
+        field = fields.Function(
+            serialize=lambda val: val.lower(),
+            deserialize=lambda val: val.upper()
+        )
+        assert field.dump_only is False
+
     def test_function_field_passed_serialize(self, user):
         field = fields.Function(serialize=lambda obj: obj.name.upper())
         assert "FOO" == field.serialize("key", user)
 
-    def test_function_field_passed_func_is_deprecated(self):
-        pytest.deprecated_call(lambda: fields.Function(func=lambda obj: obj.name.upper()))
+    # https://github.com/marshmallow-code/marshmallow/issues/395
+    def test_function_field_does_not_swallow_attribute_error(self, user):
+        def raise_error(obj):
+            raise AttributeError()
+        field = fields.Function(serialize=raise_error)
+        with pytest.raises(AttributeError):
+            field.serialize('key', user)
+
+    def test_function_field_load_only(self):
+        field = fields.Function(deserialize=lambda obj: None)
+        assert field.load_only
 
     def test_function_field_passed_serialize_with_context(self, user, monkeypatch):
         class Parent(Schema):
@@ -101,10 +121,6 @@ class TestFieldSerialization:
         user.age = None
         field = fields.Integer(default=None)
         assert field.serialize('age', user) is None
-
-    def test_callable_field(self, user):
-       field = fields.String()
-       assert field.serialize('call_me', user) == 'This was called.'
 
     def test_uuid_field(self, user):
         user.uuid1 = '{12345678-1234-5678-1234-567812345678}'
@@ -331,9 +347,9 @@ class TestFieldSerialization:
         user.falsy = 'false'
         user.none = None
 
-        assert field.serialize('truthy', user) == True
-        assert field.serialize('falsy', user) == False
-        assert field.serialize('none', user) == None
+        assert field.serialize('truthy', user) is True
+        assert field.serialize('falsy', user) is False
+        assert field.serialize('none', user) is None
 
     def test_function_with_uncallable_param(self):
         with pytest.raises(ValueError):
@@ -390,6 +406,16 @@ class TestFieldSerialization:
         with pytest.raises(ValueError):
             BadSerializer().dump(u)
 
+    def test_method_field_passed_serialize_only_is_dump_only(self, user):
+        field = fields.Method(serialize='method')
+        assert field.dump_only is True
+        assert field.load_only is False
+
+    def test_method_field_passed_deserialize_only_is_load_only(self):
+        field = fields.Method(deserialize='somemethod')
+        assert field.load_only is True
+        assert field.dump_only is False
+
     def test_method_field_with_uncallable_attribute(self):
         class BadSerializer(Schema):
             foo = 'not callable'
@@ -398,9 +424,16 @@ class TestFieldSerialization:
         with pytest.raises(ValueError):
             BadSerializer().dump(u)
 
-    def test_method_prefers_serialize_over_method_name(self):
-        m = fields.Method(serialize='serialize', method_name='method')
-        assert m.serialize_method_name == 'serialize'
+    # https://github.com/marshmallow-code/marshmallow/issues/395
+    def test_method_field_does_not_swallow_attribute_error(self):
+        class MySchema(Schema):
+            mfield = fields.Method('raise_error')
+
+            def raise_error(self, obj):
+                raise AttributeError()
+
+        with pytest.raises(AttributeError):
+            MySchema().dump({})
 
     def test_method_with_no_serialize_is_missing(self):
         m = fields.Method()
@@ -662,9 +695,9 @@ class TestFieldSerialization:
         with pytest.raises(ValidationError):
             field.serialize('dtimes', obj)
 
-    def test_list_field_work_with_generators_empty_generator_returns_none_for_every_non_returning_yield_statement(self):
+    def test_list_field_work_with_generators_empty_generator_returns_none_for_every_non_returning_yield_statement(self):  # noqa
         def custom_generator():
-            a = yield
+            yield
             yield
         obj = DateTimeList(custom_generator())
         field = fields.List(fields.DateTime, allow_none=True)